fixed prefs-function (wrong FormUtil::getPassedValue)

+++ Dizkus/pnuser.php
@@ -560,14 +560,13 @@ function Dizkus_user_prefs($args=array())
 
     // get the input
     $topic_id = (int)FormUtil::getPassedValue('topic', (isset($args['topic'])) ? $args['topic'] : null, 'GETPOST');
-    $act = (int)FormUtil::getPassedValue('act', (isset($args['act'])) ? $args['act'] : '', 'GETPOST');
+    $act =  FormUtil::getPassedValue('act', (isset($args['act'])) ? $args['act'] : '', 'GETPOST');
     $return_to = (int)FormUtil::getPassedValue('return_to', (isset($args['return_to'])) ? $args['return_to'] : '', 'GETPOST');
     $forum_id = (int)FormUtil::getPassedValue('forum', (isset($args['forum'])) ? $args['forum'] : null, 'GETPOST');
     $user_id = (int)FormUtil::getPassedValue('user', (isset($args['user'])) ? $args['user'] : null, 'GETPOST');
 
     // user_id will only be used if we have admin permissions otherwise the
     // user can edit his prefs only but not others users prefs
-
 
     switch($act) {
         case 'subscribe_topic':
